Common padding functions
